SQUASH: More refactoring
