+++ q2_sample_classifier/tests/test_classifier.py
@@ -5,19 +5,11 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
-import os
 from warnings import filterwarnings
-import tempfile
-import shutil
-import json
-import tarfile
 import pandas as pd
 import numpy as np
-import sklearn
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.feature_selection import RFECV
-from sklearn.pipeline import Pipeline
-import joblib
 import pandas.util.testing as pdt
 import biom
 
@@ -27,11 +19,7 @@ from qiime2.plugins import sample_classifier
 from q2_sample_classifier.visuals import (
     _custom_palettes, _roc_palette, _roc_per_class, _roc_micro_average,
     _roc_macro_average, _binarize_labels, _generate_roc_plots)
-from q2_sample_classifier.classify import (
-    fit_classifier, summarize)
 from q2_sample_classifier.utilities import _extract_rfe_scores
-from q2_sample_classifier import (
-    SampleEstimatorDirFmt, PickleFormat)
 from q2_sample_classifier.tests.test_base_class import \
     SampleClassifierTestPluginBase
 
@@ -123,178 +111,6 @@ class TestBinaryClassification(SampleClassifierTestPluginBase):
         exp = pd.Series([c for c in 'ababab'], name='prediction',
                         index=pd.Index([i for i in 'aebdcf'], name='id'))
         pdt.assert_series_equal(exp, res[0].view(pd.Series))
-
-
-class NowLetsTestTheActions(SampleClassifierTestPluginBase):
-
-    def setUp(self):
-        super().setUp()
-        md = pd.Series(['a', 'a', 'b', 'b', 'b'],
-                       index=['a', 'b', 'c', 'd', 'e'], name='bugs')
-        md.index.name = 'SampleID'
-        self.md = qiime2.CategoricalMetadataColumn(md)
-
-        tab = biom.Table(
-            np.array([[3, 6, 7, 3, 6], [3, 4, 5, 6, 2], [8, 6, 4, 1, 0],
-                      [8, 6, 4, 1, 0], [8, 6, 4, 1, 0]]),
-            observation_ids=['v', 'w', 'x', 'y', 'z'],
-            sample_ids=['a', 'b', 'c', 'd', 'e'])
-        self.tab = qiime2.Artifact.import_data('FeatureTable[Frequency]', tab)
-
-        md2 = pd.DataFrame({'trash': ['a', 'a', 'b', 'b', 'b', 'junk'],
-                            'floats': [0.1, 0.1, 1.3, 1.8, 1000.1, 0.1],
-                            'ints': [0, 1, 2, 2, 2, 0],
-                            'nans': [1, 1, 2, 2, np.nan, np.nan],
-                            'negatives': [-7, -3, -1.2, -4, -9, -1]},
-                           index=['a', 'b', 'c', 'd', 'e', 'peanut'])
-        md2.index.name = 'SampleID'
-        self.md2 = qiime2.Metadata(md2)
-
-    # let's make sure the correct transformers are in place! See issue 114
-    # if this runs without error, that's good enough for me. We already
-    # validate the function above.
-    def test_action_split_table(self):
-        sample_classifier.actions.split_table(self.tab, self.md, test_size=0.5)
-
-    def test_metatable(self):
-        exp = biom.Table(
-            np.array([[0.1, 0.1, 1.3, 1.8, 1000.1, 0.1],
-                      [0, 1, 2, 2, 2, 0]]),
-            observation_ids=['floats', 'ints'],
-            sample_ids=['a', 'b', 'c', 'd', 'e', 'peanut'])
-        res, = sample_classifier.actions.metatable(
-            self.md2, missing_values='drop_features')
-        report = res.view(biom.Table).descriptive_equality(exp)
-        self.assertIn('Tables appear equal', report, report)
-
-    def test_metatable_missing_error(self):
-        with self.assertRaisesRegex(ValueError, "missing values"):
-            sample_classifier.actions.metatable(
-                self.md2, missing_values='error')
-
-    def test_metatable_drop_samples(self):
-        exp = biom.Table(
-            np.array([[3, 6, 7, 3], [3, 4, 5, 6], [8, 6, 4, 1],
-                      [8, 6, 4, 1], [8, 6, 4, 1],
-                      [0.1, 0.1, 1.3, 1.8],
-                      [0, 1, 2, 2], [1, 1, 2, 2]]),
-            observation_ids=['v', 'w', 'x', 'y', 'z', 'floats', 'ints',
-                             'nans'],
-            sample_ids=['a', 'b', 'c', 'd'])
-        res, = sample_classifier.actions.metatable(
-            self.md2, self.tab, missing_values='drop_samples')
-        report = res.view(biom.Table).descriptive_equality(exp)
-        self.assertIn('Tables appear equal', report, report)
-
-    def test_metatable_fill_na(self):
-        exp = biom.Table(
-            np.array([[3, 6, 7, 3, 6], [3, 4, 5, 6, 2], [8, 6, 4, 1, 0],
-                      [8, 6, 4, 1, 0], [8, 6, 4, 1, 0],
-                      [0.1, 0.1, 1.3, 1.8, 1000.1],
-                      [0, 1, 2, 2, 2], [1., 1., 2., 2., 0.]]),
-            observation_ids=['v', 'w', 'x', 'y', 'z', 'floats', 'ints',
-                             'nans'],
-            sample_ids=['a', 'b', 'c', 'd', 'e'])
-        res, = sample_classifier.actions.metatable(
-            self.md2, self.tab, missing_values='fill')
-        report = res.view(biom.Table).descriptive_equality(exp)
-        self.assertIn('Tables appear equal', report, report)
-
-    def test_metatable_with_merge(self):
-        exp = biom.Table(
-            np.array([[3, 6, 7, 3, 6], [3, 4, 5, 6, 2], [8, 6, 4, 1, 0],
-                      [8, 6, 4, 1, 0], [8, 6, 4, 1, 0],
-                      [0.1, 0.1, 1.3, 1.8, 1000.1],
-                      [0, 1, 2, 2, 2]]),
-            observation_ids=['v', 'w', 'x', 'y', 'z', 'floats', 'ints'],
-            sample_ids=['a', 'b', 'c', 'd', 'e'])
-        res, = sample_classifier.actions.metatable(
-            self.md2, self.tab, missing_values='drop_features')
-        report = res.view(biom.Table).descriptive_equality(exp)
-        self.assertIn('Tables appear equal', report, report)
-
-    def test_metatable_with_merge_successful_inner_join(self):
-        exp = biom.Table(
-            np.array([[3, 6, 7, 3], [3, 4, 5, 6], [8, 6, 4, 1],
-                      [8, 6, 4, 1], [8, 6, 4, 1], [0.1, 0.1, 1.3, 1.8],
-                      [0, 1, 2, 2], [1., 1., 2., 2.]]),
-            observation_ids=['v', 'w', 'x', 'y', 'z', 'floats', 'ints',
-                             'nans'],
-            sample_ids=['a', 'b', 'c', 'd'])
-        res, = sample_classifier.actions.metatable(
-            self.md2.filter_ids(['a', 'b', 'c', 'd']), self.tab,
-            missing_values='error')
-        report = res.view(biom.Table).descriptive_equality(exp)
-        self.assertIn('Tables appear equal', report, report)
-
-    def test_metatable_with_merge_error_inner_join(self):
-        with self.assertRaisesRegex(ValueError, "Missing samples"):
-            sample_classifier.actions.metatable(
-                self.md2.filter_ids(['a', 'b', 'c', 'd']),
-                self.tab, missing_samples='error',
-                missing_values='drop_samples')
-
-    def test_metatable_empty_metadata_after_drop_all_unique(self):
-        with self.assertRaisesRegex(
-                ValueError, "All metadata"):  # are belong to us
-            sample_classifier.actions.metatable(
-                self.md2.filter_ids(['b', 'c']), self.tab,
-                missing_values='drop_samples', drop_all_unique=True)
-
-    def test_metatable_no_samples_after_filtering(self):
-        junk_md = pd.DataFrame(
-            {'trash': ['a', 'a', 'b', 'b', 'b', 'junk'],
-             'floats': [np.nan, np.nan, np.nan, 1.8, 1000.1, 0.1],
-             'ints': [0, 1, 2, np.nan, 2, 0],
-             'nans': [1, 1, 2, 2, np.nan, np.nan],
-             'negatives': [-7, -4, -1.2, -4, -9, -1]},
-            index=['a', 'b', 'c', 'd', 'e', 'peanut'])
-        junk_md.index.name = 'SampleID'
-        junk_md = qiime2.Metadata(junk_md)
-        with self.assertRaisesRegex(ValueError, "All metadata samples"):
-            sample_classifier.actions.metatable(
-                junk_md, missing_values='drop_samples')
-
-
-class SampleEstimatorTestBase(SampleClassifierTestPluginBase):
-    package = 'q2_sample_classifier.tests'
-
-    def setUp(self):
-        super().setUp()
-
-        def _load_biom(table_fp):
-            table_fp = self.get_data_path(table_fp)
-            table = qiime2.Artifact.load(table_fp)
-            table = table.view(biom.Table)
-            return table
-
-        def _load_cmc(md_fp, column):
-            md_fp = self.get_data_path(md_fp)
-            md = pd.read_csv(md_fp, sep='\t', header=0, index_col=0)
-            md = qiime2.CategoricalMetadataColumn(md[column])
-            return md
-
-        table_chard_fp = _load_biom('chardonnay.table.qza')
-        mdc_chard_fp = _load_cmc('chardonnay.map.txt', 'Region')
-
-        pipeline, importances = fit_classifier(
-            table_chard_fp, mdc_chard_fp, random_state=123,
-            n_estimators=2, n_jobs=1, optimize_feature_selection=True,
-            parameter_tuning=True, missing_samples='ignore')
-        transformer = self.get_transformer(
-            Pipeline, SampleEstimatorDirFmt)
-        self._sklp = transformer(pipeline)
-        sklearn_pipeline = self._sklp.sklearn_pipeline.view(PickleFormat)
-        self.sklearn_pipeline = str(sklearn_pipeline)
-        self.pipeline = pipeline
-
-    def _custom_setup(self, version):
-        with open(os.path.join(self.temp_dir.name,
-                               'sklearn_version.json'), 'w') as fh:
-            fh.write(json.dumps({'sklearn-version': version}))
-        shutil.copy(self.sklearn_pipeline, self.temp_dir.name)
-        return SampleEstimatorDirFmt(
-            self.temp_dir.name, mode='r')
 
 
 class TestROC(SampleClassifierTestPluginBase):
@@ -398,61 +214,3 @@ class TestBinarize(SampleClassifierTestPluginBase):
         exp = np.array([[1, 0, 0], [0, 1, 0], [0, 0, 1],
                         [1, 0, 0], [0, 1, 0], [0, 0, 1]])
         np.testing.assert_array_equal(exp, labels)
-
-
-class TestFormats(SampleEstimatorTestBase):
-    def test_sample_classifier_dir_fmt(self):
-        format = self._custom_setup(sklearn.__version__)
-
-        # Should not error
-        format.validate()
-
-
-class TestTransformers(SampleEstimatorTestBase):
-    def test_old_sklearn_version(self):
-        transformer = self.get_transformer(
-            SampleEstimatorDirFmt, Pipeline)
-        input = self._custom_setup('a very old version')
-        with self.assertRaises(ValueError):
-            transformer(input)
-
-    def test_taxo_class_dir_fmt_to_taxo_class_result(self):
-        input = self._custom_setup(sklearn.__version__)
-
-        transformer = self.get_transformer(
-            SampleEstimatorDirFmt, Pipeline)
-        obs = transformer(input)
-
-        self.assertTrue(obs)
-
-    def test_taxo_class_result_to_taxo_class_dir_fmt(self):
-        def read_pipeline(pipeline_filepath):
-            with tarfile.open(pipeline_filepath) as tar:
-                dirname = tempfile.mkdtemp()
-                tar.extractall(dirname)
-                pipeline = joblib.load(os.path.join(dirname,
-                                       'sklearn_pipeline.pkl'))
-                for fn in tar.getnames():
-                    os.unlink(os.path.join(dirname, fn))
-                os.rmdir(dirname)
-            return pipeline
-
-        exp = read_pipeline(self.sklearn_pipeline)
-        transformer = self.get_transformer(
-            Pipeline, SampleEstimatorDirFmt)
-        obs = transformer(exp)
-        sklearn_pipeline = obs.sklearn_pipeline.view(PickleFormat)
-        obs_pipeline = read_pipeline(str(sklearn_pipeline))
-        obs = obs_pipeline
-        self.assertTrue(obs)
-
-
-# make sure summarize visualizer works and that rfe_scores are stored properly
-class TestSummarize(SampleEstimatorTestBase):
-
-    def test_summary_with_rfecv(self):
-        summarize(self.temp_dir.name, self.pipeline)
-
-    def test_summary_without_rfecv(self):
-        del self.pipeline.rfe_scores
-        summarize(self.temp_dir.name, self.pipeline)
